fix(reports): address code review findings on best selling products
- SQL: cast quantity to bigint before multiplication to prevent
  int32 overflow on high-volume products
  (SUM(quantity::bigint * unit_price) instead of SUM(quantity * unit_price)::bigint)
- Extract fetch+state logic into useBestSellingProducts hook;
  add isCancelled flag to prevent stale fetch from overwriting
  newer state after rapid date range changes
- Remove export from BestSellingProductRow (internal cast type only)
- BestSellingProducts component now under 50 lines per CLAUDE.md

+++ supabase/migrations/20260627000001_best_selling_products_rpc.sql
@@ -18,7 +18,7 @@ AS $$
   SELECT
     oi.product_name,
     SUM(oi.quantity)::bigint                           AS quantity,
-    SUM(oi.quantity * oi.unit_price)::bigint           AS revenue
+    SUM(oi.quantity::bigint * oi.unit_price)::bigint    AS revenue
   FROM   public.order_items oi
   JOIN   public.orders      o  ON o.id = oi.order_id
   WHERE  o.status      = 'done'
